+++ R/rec.R
@@ -13,7 +13,7 @@
 #'          function to create recode strings for grouping variables.
 #' @param as.num Logical, if \code{TRUE}, return value will be numeric, not a factor.
 #' @param var.label Optional string, to set variable label attribute for the
-#'          returned variable (see vignette \href{../doc/intro_sjmisc.html}{Labelled Data and the sjmisc-Package}).
+#'          returned variable (see vignette \href{https://cran.r-project.org/package=sjlabelled/vignettes/intro_sjlabelled.html}{Labelled Data and the sjlabelled-Package}).
 #'          If \code{NULL} (default), variable label attribute of \code{x} will
 #'          be used (if present). If empty, variable label attributes will be removed.
 #' @param val.labels Optional character vector, to set value label attributes